+++ qom/loopers/XLooper.py
@@ -6,14 +6,14 @@
 __name__    = 'qom.loopers.XLooper'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-12-21'
-__updated__ = '2021-01-06'
+__updated__ = '2021-01-11'
 
 # dependencies
 import logging
 import numpy as np
 
 # qom modules
-from qom.loopers.BaseLooper import BaseLooper
+from .BaseLooper import BaseLooper
 
 # module logger
 logger = logging.getLogger(__name__)