+++ qom/ui/plotters/__init__.py
@@ -1,1 +1,1 @@
-from qom.ui.plotters.MPLPlotter import MPLPlotter
+from .MPLPlotter import MPLPlotter